pyproject.toml
[project]
name = "discos"
version = "1.0.0"
description = "DiscOS: Discovery OS userland for EvidenceOS (UVP-compatible)."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "DiscOS Contributors"}]

dependencies = [
  "pydantic>=2.0",
  "jsonschema>=4.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
wasm = ["wasmtime>=16.0.0"]
dev = ["pytest>=7.0", "ruff>=0.4", "mypy>=1.0"]

[project.scripts]
discos = "discos.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = ["src"]
